Remove more implicit casts in upnp part
Remove more "implicit integer or enum conversions" as well as memset
before snprintf.

(forward port of commit 2eb3e069badd5c8676738c3ead37f9551fd8448e)

+++ upnp/src/api/upnpapi.c
@@ -1313,9 +1313,9 @@ int UpnpUnRegisterRootDeviceLowPower(UpnpDevice_Handle Hnd, int PowerState,
 	if (HInfo->aliasInstalled)
 		web_server_set_alias(NULL, NULL, 0, 0);
 #endif /* INTERNAL_WEB_SERVER */
-	if (HInfo->DeviceAf == AF_INET)
+	if (HInfo->DeviceAf == (unsigned short)AF_INET)
 		UpnpSdkDeviceRegisteredV4 = 0;
-	else if (HInfo->DeviceAf == AF_INET6)
+	else if (HInfo->DeviceAf == (unsigned short)AF_INET6)
 		UpnpSdkDeviceregisteredV6 = 0;
 	FreeHandle(Hnd);
 	HandleUnlock();
